refactor day two functor project
