+++ first-week/day-two/typings/functor.ts
@@ -1,0 +1,63 @@
+/**
+ * LUXCIUM LICENSE *NO* PERMISSION GRANTED - PROVIDED "AS IS" - WITHOUT WARRANTY
+ * =============================================================================≈
+ * @author Benjamin Vincent Kasapoglu (LUXCIUM)
+ * @copyright © 2020 - LUXCIUM (Benjamin Vincent Kasapoglu) <luxcium@neb401.com>
+ * @license LUXCIUM-LICENSE for this file only or other having the same mention
+ * =============================================================================≈
+ * Scientia es lux principium✨ ™ - SEE THE BOTTOM OF THIS FILES FOR MORE INFO
+ */
+
+/** A Map takes any value A and morph it into any value B */
+export type MapType<A = unknown> = (fn: <B = unknown>(val: A) => B) => IFMap;
+/**
+ * Anything that implement a Map must implement
+ * a method of the form map: MapType<A>
+ */
+export interface IFMap<A = unknown> {
+  /** `fantasy-land/map :: Functor f => f a ~> (a -> b) -> f b` */
+  map: MapType<A>;
+}
+
+export interface IFork<A = unknown> {
+  // extends ValueType<A>
+  /** Return the internal value of a Functor or type extending Functor */
+  readonly fork: A;
+}
+
+/** A complex Functor must extend simple Functor and fork */
+
+// export function testing() {
+
+/*
+================================================================================≈
+† LUXCIUM‡ LICENSE — *NO* PERMISSION GRANTED — PROVIDED "AS IS" WITHOUT WARRANTY
+
+‽  *NO* PERMISSION ARE GRANTED. NOT TO PUBLISH, NOT TO DISTRIBUTE, NOT TO
+‽  SUBLICENSE, AND/OR NOT TO SELL COPIES OF THE SOFTWARE.
+‽
+‽  You may use, only for yourself, the source code in this file and from
+‽  imported files to this one. You have the right to: distribute to your
+‽  friends, to your students, or to you coworker for personal or academic use
+‽  at home at school or at work. You must include this «FULL NOTICE» in each
+‽  imported files to this one that you decide to use if such «FULL NOTICE»
+‽  is missing. If you modify the code in a substatial way you may decide to
+‽  include only the «COPYRIGHT» and the «SHORT NOTICE» below.
+
+† Copyright © 2020 — LUXCIUM‡ (Benjamin Vincent Kasapoglu) <luxcium⸓neb401.com>
+
+- --> SHORT NOTICE
+-  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ALL KIND, EXPRESS OR
+-  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+-  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+-  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ALL CLAIM, DAMAGES OR OTHER
+-  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+-  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+-  SOFTWARE.
+- -->
+-
+- IN ANY CASES THE COPYRIGHT AND SHORT NOTICE ABOVE MUST BE INCLUDED.
+
+† Scientia es lux principium✨ ™
+================================================================================≈
+*/
